WIP: Generic Data Bundles
  * Split up inheritance structure of Bundle classes
